refactor: organização das pastas de tests
