+++ backend/plataform/tests/testsCliente/testsClienteIntegracao.py
@@ -1,0 +1,29 @@
+from plataform.models import Cliente
+from rest_framework.test import APITestCase
+
+
+class ClienteTestsIntegracao(APITestCase):
+    def setUp(self):
+        data = {
+            "username": "MathewsDantas",
+            "first_name": "Mathews",
+            "last_name": "Santos",
+            "email": "cliente@hotmail.com",
+            "cpf": "11945011614",
+            "password": "Tads.d.b.s123",
+        }
+        self.cliente = Cliente.objects.create_user(**data)
+
+    def test_update_cliente(self):
+        cliente = Cliente.objects.get(email="cliente@hotmail.com")
+        cliente.email = "cliente2@hotmail.com"
+        cliente.save()
+
+        updated_cliente = Cliente.objects.get(id=self.cliente.id)
+        self.assertEqual(updated_cliente.email, "cliente2@hotmail.com")
+
+    def test_delete_cliente(self):
+        cliente = Cliente.objects.get(email="cliente@hotmail.com")
+        cliente.delete()
+
+        self.assertEqual(Cliente.objects.filter(email="cliente@hotmail.com").exists(), False)
